A couple of chr_id param type changes
In `MsGetRndChr`:

```
        0061adea 8b 7d 08        MOV      EDI,dword ptr [EBP + Stack[0x4]]
        0061aded 57                 PUSH      EDI
        0061adee e8 fd ad        CALL       MsGetRamChrMonster
```
If I've understood correctly, the instruction at `61adea` is reading a 4 byte value, which is `param_1` for these functions.

+++ src/core/ffx2/call.cs
@@ -186,7 +186,7 @@ public static unsafe partial class FhCall {
         => new( new FhMethodLocation("FFX-2.exe", 0x218B80) ) ;
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate int d_MsGetRndChr(int chr_id, int arg2);
+    public delegate int d_MsGetRndChr(uint chr_id, int arg2);
     public static FhMethodHandle<d_MsGetRndChr> MsGetRndChr 
         =>new( new FhMethodLocation("FFX-2.exe", 0x21ADD0) );
 
@@ -231,7 +231,7 @@ public static unsafe partial class FhCall {
         => new( new FhMethodLocation("FFX-2.exe", 0x224F90) );
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate uint d_MsGetRamChrMonster(byte chr_id);
+    public delegate uint d_MsGetRamChrMonster(uint chr_id);
     public static FhMethodHandle<d_MsGetRamChrMonster> MsGetRamChrMonster 
         => new( new FhMethodLocation("FFX-2.exe", 0x225BF0) );
